+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,7 +1,11 @@
+## 2.4.15+1
+
+* [Supports Impeller](https://docs.flutter.dev/release/breaking-changes/android-surface-plugins)
+
 ## 2.4.15
 
-* Updates minimum supported SDK version to Flutter 3.22/Dart 3.2.
-* [Supports Impeller](https://docs.flutter.dev/release/breaking-changes/android-surface-plugins)
+* Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
+* Removes support for apps using the v1 Android embedding.
 
 ## 2.4.14